requirements.txt
argparse~=1.4.0
boto3~=1.34.96
botocore~=1.34.96
chromadb
flashrank~=0.2.4
ipython~=8.24.0
jupyter
langchain~=0.1.17
langchain-community
langchain-openai
langchain-aws
langchainhub
Markdown
pandas~=2.2.2
python-dotenv~=1.0.1
tiktoken~=0.6.0
tqdm
unstructured
unstructured-client
